prevent provider and view from being hoisted up into sentry.auth, and fix imports
