+++ src/sentry/auth/__init__.py
@@ -1,8 +1,6 @@
 from __future__ import absolute_import, print_function
 
-from .provider import *  # NOQA
 from .manager import ProviderManager
-from .view import *  # NOQA
 
 manager = ProviderManager()
 register = manager.register
