+++ src/sentry/auth/providers/dummy.py
@@ -2,8 +2,8 @@ from __future__ import absolute_import, print_function
 
 from django.http import HttpResponse
 
-from sentry.auth import Provider, AuthView
-from sentry.auth.provider import MigratingIdentityId
+from sentry.auth.provider import Provider, MigratingIdentityId
+from sentry.auth.view import AuthView
 
 
 class AskEmail(AuthView):
